+++ websift/__init__.py
@@ -20,7 +20,7 @@ Usage:
     )
 """
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"
 
 from websift.client import WebSearchClient
 from websift.settings import AppSettings